% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.Grid.R
\docType{class}
\name{Coverage.Grid-class}
\alias{Coverage.Grid-class}
\title{Class "Coverage.Grid"}
\description{
Class \code{"Coverage.Grid"} is an S4 class containing descriptions of a
grid used to assess the coverage scores of different designs.
}
\section{Slots}{

\describe{
\item{\code{grid}}{sf multipoint object}

\item{\code{spacing}}{the spacing used to create the coverage grid}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{make.grid(region = make.region(), no.points = 1000, spacing = numeric(0)}
}

\section{Methods}{

\describe{
 \item{\code{plot}}{\code{signature(x = "Coverage.Grid", y = "missing")}: plots
 the grid of points.}
}
}

\keyword{classes}
