% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R, R/Survey.Design.R
\docType{methods}
\name{get.coverage}
\alias{get.coverage}
\alias{get.coverage,Survey.Design-method}
\title{S4 generic method to extract coverage scores}
\usage{
get.coverage(object)

\S4method{get.coverage}{Survey.Design}(object)
}
\arguments{
\item{object}{an object which inherits from class Survey.Design}
}
\value{
a vector of coverage scores
}
\description{
Obtains the coverage scores from the survey design object.
}
