% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line.Transect.Design.R
\docType{class}
\name{Line.Transect.Design-class}
\alias{Line.Transect.Design-class}
\title{Virtual Class "Line.Transect.Design" extends Class "Survey.Design"}
\description{
Virtual Class \code{"Line.Transect.Design"} is an S4 class detailing
the type of line transect design.
}
\section{Slots}{

\describe{
\item{\code{design}}{Character value describing the name of the design.}

\item{\code{line.length}}{Numeric value defining the total line length to be generated (may be
multiple values relating to each stratum).}

\item{\code{spacing}}{Numeric value defining the systematic spacing (may be multiple values
relating to each stratum).}
}}

\section{Methods}{

\describe{
 \item{\code{generate.transects}}{\code{signature=(object = "Line.Transect.Design", ...)}:
 generates a set of transects from a shapefile.}
}
}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
