% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line.Transect.Survey.R
\docType{class}
\name{Line.Transect.Survey-class}
\alias{Line.Transect.Survey-class}
\title{Class "Line.Transect.Survey" extends Class "Survey"}
\description{
Virtual Class \code{"Line.Transect.Survey"} is an S4 class
detailing a set of transects from a point transect design.
}
\section{Slots}{

\describe{
\item{\code{line.length}}{the total line length for the transect set}
}}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
