% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.Constructors.R
\name{make.region}
\alias{make.region}
\title{Creates a Region object}
\usage{
make.region(region.name = "region", strata.name = character(0),
  shape = NULL)
}
\arguments{
\item{region.name}{the region name}

\item{strata.name}{the stratum names (character vector, same length as the
number of areas in the \code{shapefile} or \code{coords} arguments). If not supplied "A", "B", "C", ... will be assigned.}

\item{shape}{sf object, sp object, shapefile path to .shp file or a
data.frame of coordinates.}

\item{units}{measurement units; either \code{"m"} for metres or \code{"km"} for
kilometres.}
}
\value{
object of class Region
}
\description{
This creates an instance of the Region class. If the
\code{shapefile} argument is supplied, all information will be extracted from
there. Otherwise, the a list of polygons describing the areas of interest needs
to be supplied (\code{coords}) and optionally a list of polygons describing the
areas to be excluded (\code{gaps}). If \code{area} is not specified it will be
calculated.
}
\examples{
# A basic study region of 2000m by 500m is created using the defaults
region <- make.region()
plot(region)

# Here is an example of a 1000 x 1000 study region with a gap
coords <- gaps <- list()
coords[[1]] <- list(data.frame(x = c(0,1000,1000,0,0), y = c(0,0,
 1000,1000,0)))
gaps[[1]] <- list(data.frame(x = c(400,600,500,350,400), y = c(100,
 250,600,120,100)))

region <- make.region(region.name = "study.area", units = "m",
 coords = coords, gaps = gaps)
plot(region)

}
\author{
Laura Marshall
}
