% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.Grid.R
\name{plot,Coverage.Grid,ANY-method}
\alias{plot,Coverage.Grid,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Coverage.Grid,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class Coverage.Grid}

\item{y}{not used}

\item{...}{other general plot parameters including: add as TRUE / FALSE, col as
a colour for grid points and pch as grid point symbols.}
}
\description{
Plots an S4 object of class 'Coverage.Grid'
}
