% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{plot,Region,ANY-method}
\alias{plot,Region,ANY-method}
\alias{plot,Region,Transect-method}
\title{Plot}
\usage{
\S4method{plot}{Region,ANY}(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  line.col = gray(0.2),
  legend.params = list()
)

\S4method{plot}{Region,Transect}(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  line.col = "blue",
  covered.area = FALSE,
  legend.params = list(),
  ...
)
}
\arguments{
\item{x}{object of class Region or inheriting from Survey}

\item{y}{optionally a Survey object to plot with the Region}

\item{main}{the main title for the plot}

\item{region.col}{colours for the strata}

\item{strata}{the strata name or number to be plotted. By default
all strata will be plotted.}

\item{line.col}{sets the line colour for the transects}

\item{legend.params}{depricated since implementation of ggplot2}

\item{covered.area}{boolean value saying whether the covered area should be plotted.}
}
\value{
returns a ggplot object

returns a ggplot object
}
\description{
Plot

Plot
}
