% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.Transect.Survey.R
\docType{methods}
\name{plot,Point.Transect.Survey,ANY-method}
\alias{plot,Point.Transect.Survey,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Point.Transect.Survey,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class Survey}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Survey'
}
