% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.Design.R
\docType{methods}
\name{plot,Survey.Design,ANY-method}
\alias{plot,Survey.Design,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Survey.Design,ANY}(x, y, strata.id = numeric(0),
  col.breaks = 10, subtitle = "", ...)
}
\arguments{
\item{x}{object of class Survey.Design}

\item{y}{not used}

\item{strata.id}{a numeric value indicating the index of the strata you wish to plot.}

\item{col.breaks}{the number of break point in the colour scale representing the
coverage scores.}

\item{subtitle}{a subtitle for the plot.}

\item{...}{not implemented for this class.}
}
\description{
Plots the coverage scores contained within an object of class 'Survey.Design' and
provides a colour key relating to the coverage scores. This allows the user to
assess how even the coverage is across the survey region.
}
