% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.Transect.R
\docType{class}
\name{Point.Transect-class}
\alias{Point.Transect-class}
\title{Class "Point.Transect" extends Class "Survey"}
\description{
Virtual Class \code{"Point.Transect"} is an S4 class
detailing a set of transects from a point transect design.
}
\seealso{
\code{\link{make.design}}
}
\keyword{classes}
