% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.Transect.Design.R
\docType{class}
\name{Point.Transect.Design-class}
\alias{Point.Transect.Design-class}
\title{Virtual Class "Point.Transect.Design" extends Class "Survey.Design"}
\description{
Virtual Class \code{"Point.Transect.Design"} is an S4 class detailing
the type of point transect design.
}
\section{Methods}{

\describe{
 \item{\code{generate.transects}}{\code{signature=(object = "Point.Transect.Design", quiet = FALSE,  ...)}:
 generates a set of transects from the design.}
}
}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
