% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line.Transect.R, R/Point.Transect.R
\name{show,Line.Transect-method}
\alias{show,Line.Transect-method}
\alias{show,Point.Transect-method}
\title{Show}
\usage{
\S4method{show}{Line.Transect}(object)

\S4method{show}{Point.Transect}(object)
}
\arguments{
\item{object}{an object of class Transect}
}
\description{
Displays details of an S4 object of class 'Transect'

Displays details of an S4 object of class 'Transect'
}
