% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.Design.R
\name{summary,Survey.Design-method}
\alias{summary,Survey.Design-method}
\title{summary}
\usage{
\S4method{summary}{Survey.Design}(object)
}
\arguments{
\item{object}{an object which inherits from the Survey.Design class}
}
\description{
Summarises S4 object of class 'Survey.Design'
}
