% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.Design.R
\docType{class}
\name{Survey.Design-class}
\alias{Survey.Design-class}
\title{S4 Class "Survey.Design"}
\description{
Virtual Class "Survey.Design"
}
\details{
Virtual Class \code{"Survey.Design"} is an S4 class detailing the survey
design.
}
\section{Slots}{

\describe{
\item{\code{truncation}}{Object of class \code{"numeric"}; The maximum distance
at which observations can be made. This is used to determine the covered
area during the coverage calculations.}

\item{\code{no.samplers}}{Numeric values defining the number of samplers in each
stratum.}

\item{\code{edge.protocol}}{Character value defining whether a "minus" or "plus"
sampling strategy should be used.}
}}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
