% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R, R/Line.Transect.Design.R,
%   R/Point.Transect.Design.R
\docType{methods}
\name{generate.transects}
\alias{generate.transects}
\alias{generate.transects,Line.Transect.Design-method}
\alias{generate.transects,Point.Transect.Design-method}
\title{S4 generic method to generate an instance of a design}
\usage{
generate.transects(object, silent, ...)

\S4method{generate.transects}{Line.Transect.Design}(object,
  silent = FALSE)

\S4method{generate.transects}{Point.Transect.Design}(object,
  silent = FALSE)
}
\arguments{
\item{object}{an object which inherits from class Survey.Design}

\item{silent}{silences some warnings}

\item{...}{optional argument index if an object of class Survey.Design is
supplied allowing the user to access / plot different sets of transects
listed in the filenames slot.}
}
\value{
an object of class Line.Transect
}
\description{
Uses the Survey.Design details to generate transects. Currently this
involves loading a survey shapefile from the path specified in the
Survey.Design object and can only work with line transect designs.
}
