% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R, R/Line.Transect.Design.R,
%   R/Point.Transect.Design.R
\docType{methods}
\name{generate.transects}
\alias{generate.transects}
\alias{generate.transects,Line.Transect.Design-method}
\alias{generate.transects,Point.Transect.Design-method}
\title{S4 generic method to generate an instance of a design}
\usage{
generate.transects(object, quiet = FALSE, ...)

\S4method{generate.transects}{Line.Transect.Design}(object,
  quiet = FALSE, ...)

\S4method{generate.transects}{Point.Transect.Design}(object,
  quiet = FALSE, ...)
}
\arguments{
\item{object}{an object which inherits from class Survey.Design}

\item{quiet}{silences some warnings}

\item{...}{optional arguments used for internal calls}
}
\value{
an object of class Transect
}
\description{
Uses the Survey.Design details to generate a set of transects.
The transects are returned within an object of class Transect which records
some of the design options used to generate it along with the samplers as
an sf object of class 'POINT' or 'LINESTRING'/'MULTILINESTRING'. The
Transect object also contains the covered areas as a 'POLYGON' or
'MULTIPOLYGON' sf object.
}
