% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R, R/Survey.Design.R
\name{get.coverage}
\alias{get.coverage}
\alias{get.coverage,Survey.Design-method}
\title{S4 generic method to extract coverage scores}
\usage{
get.coverage(object, strata.id = "all")

\S4method{get.coverage}{Survey.Design}(object, strata.id = "all")
}
\arguments{
\item{object}{an object which inherits from class Survey.Design}

\item{strata.id}{either "all" or a numeric value indicating the strata index.}
}
\value{
a vector of coverage scores
}
\description{
Obtains the coverage scores from the survey design object.
}
\details{
See ?make.design for example code
}
\seealso{
\link{make.design}
}
