% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line.Transect.R
\docType{class}
\name{Line.Transect-class}
\alias{Line.Transect-class}
\title{Class "Line.Transect" extends Class "Transect"}
\description{
Virtual Class \code{"Line.Transect"} is an S4 class
detailing a set of transects from a point transect design.
}
\section{Slots}{

\describe{
\item{\code{line.length}}{the total line length for the transect set}

\item{\code{trackline}}{the total on and off effort trackline length from
the start of the first transect to the end of the last}

\item{\code{cyclictrackline}}{the trackline distance plus the distance
required to return from the end of the last transect to the
beginning of the first}
}}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
