% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.Constructors.R
\name{make.coverage}
\alias{make.coverage}
\title{Creates a Coverage.Grid object}
\usage{
make.coverage(region = make.region(), spacing = numeric(0),
  n.grid.points = 1000)
}
\arguments{
\item{region}{the region name}

\item{spacing}{spacing to be used to create the coverage grid}

\item{n.grid.points}{the desired number of grid points (note that
the exact number generated may differ slightly depending on the
shape of the study region).}
}
\value{
object of class Coverage.Grid
}
\description{
This creates an instance of the Coverage.Grid class.
}
\author{
Laura Marshall
}
