% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.Constructors.R
\name{make.coverage}
\alias{make.coverage}
\title{Creates a Coverage.Grid object}
\usage{
make.coverage(region = make.region(), spacing = numeric(0),
  n.grid.points = 1000, grid = list())
}
\arguments{
\item{region}{the region name}

\item{spacing}{the stratum names (character vector, same length as the
number of areas in the \code{shapefile} or \code{coords} arguments). If not supplied "A", "B", "C", ... will be assigned.}

\item{grid}{measurement units; either \code{"m"} for metres or \code{"km"} for
kilometres.}
}
\value{
object of class Coverage.Grid
}
\description{
This creates an instance of the Coverage.Grid class.
}
\examples{
# A coverage grid in a rectangular region of 2000 x 500
region <- make.coverage(make.region(), spacing = 10)
plot(region, grid)
}
\author{
Laura Marshall
}
