% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.Constructors.R
\name{make.design}
\alias{make.design}
\title{Creates a Survey.Design object}
\usage{
make.design(region = make.region(), transect.type = "line",
  design = "systematic", no.samplers = numeric(0),
  line.length = numeric(0), effort.allocation = numeric(0),
  design.angle = 0, spacing = numeric(0), edge.protocol = "minus",
  bounding.shape = "rectangle", truncation = 1, coverage.grid = NULL)
}
\arguments{
\item{region}{an object of class Region defining the survey region.}

\item{transect.type}{character variable specifying either "line" or "point"}

\item{design}{a character variable describing the type of design. Either "random",
"systematic" or "ESzigzag" (equal-spaced zigzag). See details for more information.}

\item{no.samplers}{the number of samplers you wish the design to generate.}

\item{line.length}{the total line length you desire.}

\item{effort.allocation}{numeric values used to indicate the proportion of effort
to be allocated to each strata from number of samplers or line length. If length 0,
effort allocated based on stratum area.}

\item{design.angle}{numeric value detailing the angle of the design. Can provide
multiple values relating to strata. The use of the angle varies with design, it
can be either the angle of the grid of points, the angle of lines or the design
axis for the zigzag design.}

\item{spacing}{used by systematic designs, numeric value to define spacing
between transects.}

\item{edge.protocol}{character value indicating whether a "plus" sampling or
"minum" sampling protocol is used.}

\item{bounding.shape}{only applicable to zigzag designs. A character value saying
whether the zigzag transects should be generated using a minimum bounding
"rectangle" or a "convex hull".}

\item{truncation}{A numeric value describing the longest distance at which an
object may be observed.}
}
\value{
object of a class which inherits from class Survey.Design
}
\description{
Creates a description of a survey design. Designs may use different
types of either point or line transect designs across strata but cannot mix
point and line transect design types within a single design object.
}
\details{
For point transect designs the user may either specify "random" or
"systematic" for the design argument. If the user specifies "random", they
should also provide a value for effort detailing the number of point transects
they wish their survey to have (for stratified designs they may specify a vector
of numbers detailing the number of transects per strata or alternatively use the
effort.allocation argument to allocate effort proportionally). If the user specified
"systematic" they may either provide their desired number of samplers or a value
for spacing which defines the gap between each of the points (again a vector
of spacing values can be provided for each strata). Optionally the user may
select a design.angle. For both random and systematic point transect designs
the user may select either a minus or plus sampling edge protocol.

For line transect designs the user may either specify "random" (randomly
placed full width lines), "systematic" (systematically placed full width lines)
or "ESzigzag" (equally placed zigzag lines). If the user specifies "random", they
should provide the either the number of samplers they wish the design to generate
or the total line length they wish to achieve. If the user specifies "systematic"
they should specify either the number of samplers, the desired total line length
or the spacing between lines. The design angle for these parallel line designs
refers to the angle of the lines. If the user specifies the zigzag design they
should specify the systematic spacing value to be used and should choose between
generating the design in a minimum bounding rectangle or a convex hull. The
designs may be generated using plus or minus sampling protocols. Similar to
the point transect designs different values may be specified for each strata
for all of the above options.
}
\examples{
design <- make.design(transect.type = "point", no.samplers = 25, design.angle = 45)
}
\author{
Laura Marshall
}
