% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.Constructors.R
\name{make.region}
\alias{make.region}
\title{Creates a Region object}
\usage{
make.region(region.name = "region", strata.name = character(0),
  units = "m", shape = NULL)
}
\arguments{
\item{region.name}{the region name}

\item{strata.name}{the stratum names (character vector, same length as the
number of areas in the \code{shapefile} or \code{coords} arguments). If not supplied "A", "B", "C", ... will be assigned.}

\item{units}{measurement units; either \code{"m"} for metres or \code{"km"} for
kilometres.}

\item{shape}{sf object, sp object, shapefile path to .shp file or a
data.frame of coordinates.}
}
\value{
object of class Region
}
\description{
This creates an instance of the Region class. If the
\code{shapefile} argument is supplied, all information will be extracted from
there. Otherwise, the a list of polygons describing the areas of interest needs
to be supplied (\code{coords}) and optionally a list of polygons describing the
areas to be excluded (\code{gaps}). If \code{area} is not specified it will be
calculated.
}
\examples{
# A basic study region with two strata
region <- make.region()
plot(region)
}
\author{
Laura Marshall
}
