% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.Constructors.R
\name{make.region}
\alias{make.region}
\title{Creates a Region object}
\usage{
make.region(region.name = "region", strata.name = character(0),
  units = character(0), shape = NULL)
}
\arguments{
\item{region.name}{the region name}

\item{strata.name}{the stratum names (character vector, same length as the
number of areas in the shapefile / sf object). If not supplied "A", "B",
"C", ... will be assigned.}

\item{units}{measurement units; either \code{"m"} for metres or \code{"km"} for
kilometres. If the shapefile has a projection file associated with it the unit
will be taken from there.}

\item{shape}{shapefile path to .shp file or an sf object of class sf, sfc or sfg.}
}
\value{
object of class Region
}
\description{
This creates an instance of the Region class which defines the study
area for the survey.
}
\examples{
# A basic study rectangular study region
region <- make.region()
plot(region)

#Load the region from shapefile
shapefile.name <- system.file("extdata", "TrackExample.shp", package = "dssd")
region <- make.region(region.name = "study area",
                     shape = shapefile.name)
plot(region)
}
\author{
Laura Marshall
}
