% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{plot,Region,ANY-method}
\alias{plot,Region,ANY-method}
\alias{plot,Region,Transect-method}
\alias{plot,Region,Coverage.Grid-method}
\title{Plot}
\usage{
\S4method{plot}{Region,ANY}(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  line.col = gray(0.2),
  legend.params = list()
)

\S4method{plot}{Region,Transect}(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  line.col = gray(0.2),
  col = "blue",
  lwd = 1,
  covered.area = FALSE,
  legend.params = list(),
  ...
)

\S4method{plot}{Region,Coverage.Grid}(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  line.col = gray(0.2),
  col = "black",
  cex = 1,
  ...
)
}
\arguments{
\item{x}{object of class Region or inheriting from Survey}

\item{y}{an object inheriting from class Transect or an object of class
Coverage.Grid}

\item{main}{the main title for the plot}

\item{region.col}{colours for the strata}

\item{strata}{the strata name or number to be plotted. By default
all strata will be plotted.}

\item{line.col}{sets the line colour for the lines around the survey region.}

\item{legend.params}{depricated since implementation of ggplot2}

\item{col}{sets the colour of the transects / coverage grid points}

\item{lwd}{sets the line width of the transects}

\item{covered.area}{boolean value saying whether the covered area should be plotted.}

\item{cex}{affects the size of the point on the coverage grid}
}
\value{
returns a ggplot object
}
\description{
Plot's an object of class Region with optionally a set of transects or
the points of a coverage grid if supplied as the y argument.
}
