% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{plot,Region,ANY-method}
\alias{plot,Region,ANY-method}
\alias{plot,Region,Transect-method}
\alias{plot,Region,Coverage.Grid-method}
\title{Plot}
\usage{
\S4method{plot}{Region,ANY}(
  x,
  y,
  main = "",
  region.col = "default",
  strata = "all",
  scale = 1,
  line.col = gray(0.2),
  legend.params = list()
)

\S4method{plot}{Region,Transect}(
  x,
  y,
  main = "",
  region.col = "default",
  subtitle = "",
  covered.area = FALSE,
  ...
)

\S4method{plot}{Region,Coverage.Grid}(x, y, main = "", region.col = "default", subtitle = "", ...)
}
\arguments{
\item{x}{object of class Region or inheriting from Survey}

\item{y}{optionally a Survey object to plot with the Region}

\item{main}{the main title for the plot}

\item{region.col}{colours for the strata}

\item{strata}{the strata name or number to be plotted. By default
all strata will be plotted.}

\item{scale}{used to scale the x and y values in the plot (warning may give
unstable results when a projection is defined for the study area!)}

\item{line.col}{sets the line colour for the shapefile}

\item{legend.params}{depricated since implementation of ggplot2}

\item{subtitle}{a subtitle for the plot}

\item{covered.area}{boolean value saying whether the covered area should be plotted.}

\item{...}{Additional plot arguments passed to the plot method for the y argument.}
}
\description{
Plots an S4 object of class 'Region'
}
