% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{plot,Region,ANY-method}
\alias{plot,Region,ANY-method}
\alias{plot,Region,Transect-method}
\alias{plot,Region,Coverage.Grid-method}
\title{Plot}
\usage{
\S4method{plot}{Region,ANY}(
  x,
  y,
  main = "",
  region.col = "default",
  legend.params = list(inset = c(-0.2, 0), cex = 0.75, wrap = 15),
  ...
)

\S4method{plot}{Region,Transect}(
  x,
  y,
  main = "",
  region.col = "default",
  subtitle = "",
  covered.area = FALSE,
  ...
)

\S4method{plot}{Region,Coverage.Grid}(x, y, main = "", region.col = "default", subtitle = "", ...)
}
\arguments{
\item{x}{object of class Region or inheriting from Survey}

\item{y}{optionally a Survey object to plot with the Region}

\item{main}{the main title for the plot}

\item{region.col}{colours for the strata}

\item{legend.params}{a list of parameters which affect the location and appearance
of the legend. 'inset' affects the location of the legend, 'cex' affects the text
size and 'wrap' is the number of character in a line before the text is wrapped on
to the next line.}

\item{...}{Additional plot arguments passed to the plot method for the y argument.}

\item{subtitle}{a subtitle for the plot}

\item{covered.area}{boolean value saying whether the covered area should be plotted.}
}
\description{
Plots an S4 object of class 'Region'
}
