% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{plot,Region,ANY-method}
\alias{plot,Region,ANY-method}
\alias{plot,Region,Survey-method}
\title{Plot}
\usage{
\S4method{plot}{Region,ANY}(x, y, main = "", cols = c(2, 4, 5, 6, 7, 8,
  3), ...)

\S4method{plot}{Region,Survey}(x, y, main = "", region.col = c(2, 4, 5,
  6, 7, 8, 3), ...)
}
\arguments{
\item{x}{object of class Region or inheriting from Survey}

\item{y}{optionally a Survey object to plot with the Region}

\item{main}{the main title for the plot}

\item{cols}{colours for the strata}

\item{...}{other general plot parameters}

\item{region.cols}{fill colours for strata}
}
\description{
Plots an S4 object of class 'Region'

Plots an S4 object of class 'Region'
}
