% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{plot,Region,ANY-method}
\alias{plot,Region,ANY-method}
\alias{plot,Region,Survey-method}
\title{Plot}
\usage{
\S4method{plot}{Region,ANY}(x, y, main = "", cols = c(2, 4, 5, 6, 7, 8,
  3), ...)

\S4method{plot}{Region,Survey}(x, y, main = "", region.cols = c(2, 4,
  5, 6, 7, 8, 3), ...)
}
\arguments{
\item{x}{object of class Region}

\item{y}{not used}

\item{main}{character plot title}

\item{...}{other general plot parameters}

\item{x}{object of class Region}

\item{y}{object which inherits from class Survey}

\item{main}{character plot title}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Region'

Plots an S4 object of class 'Region'
}
