% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{plot,Region-method}
\alias{plot,Region-method}
\title{Plot}
\usage{
\S4method{plot}{Region}(x, y, main = "", ...)
}
\arguments{
\item{x}{object of class Region}

\item{y}{not used}

\item{main}{character plot title}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Region'
}
