% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.Grid.R, R/Line.Transect.Survey.R,
%   R/Point.Transect.Survey.R
\docType{methods}
\name{plot,Coverage.Grid,ANY-method}
\alias{plot,Coverage.Grid,ANY-method}
\alias{plot,Line.Transect.Survey,ANY-method}
\alias{plot,Point.Transect.Survey,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Coverage.Grid,ANY}(x, y, ...)

\S4method{plot}{Line.Transect.Survey,ANY}(x, y, ...)

\S4method{plot}{Point.Transect.Survey,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class Survey}

\item{y}{not used}

\item{...}{other general plot parameters}

\item{x}{object of class Survey}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Survey'

Plots an S4 object of class 'Survey'

Plots an S4 object of class 'Survey'
}
