% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.Design.R
\docType{methods}
\name{plot,Survey.Design,ANY-method}
\alias{plot,Survey.Design,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Survey.Design,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class Survey.Design}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Survey.Design'
}
