% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.Design.R
\docType{methods}
\name{plot,Survey.Design,ANY-method}
\alias{plot,Survey.Design,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Survey.Design,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class Survey.Design}

\item{y}{not used}

\item{...}{other general plot parameters, strata.id for specifying the numeric index
of a single strata to plot, subtitle for adding a subtitle to the plot and col.breaks
for specifying the number of colour break points in the colour scale.}
}
\description{
Plots the coverage scores contained within an object of class 'Survey.Design' and
provides a colour key relating to the coverage scores. This allows the user to
assess how even the coverage is across the survey region.
}
