% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line.Transect.R, R/Point.Transect.R
\docType{methods}
\name{plot,Line.Transect,ANY-method}
\alias{plot,Line.Transect,ANY-method}
\alias{plot,Point.Transect,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Line.Transect,ANY}(x, y, ...)

\S4method{plot}{Point.Transect,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class Transect}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Transect'

Plots an S4 object of class 'Transect'
}
