% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.coverage.R
\name{run.coverage}
\alias{run.coverage}
\title{run.coverage}
\usage{
run.coverage(design, reps = 10)
}
\arguments{
\item{design}{an object which inherits from the Survey.Design
class.}

\item{reps}{the number of times you wish the coverage simulation
to be carried out.}
}
\description{
This function can be used to assess the coverage of
a design. It generates many sets of transects from the design
supplied and counts how many times each point in the coverage
grid is sampled on each survey. These 'hits' are then averaged
across the number of repetitions.
}
