% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.Design.R
\docType{methods}
\name{show,Survey.Design-method}
\alias{show,Survey.Design-method}
\title{show}
\usage{
\S4method{show}{Survey.Design}(object)
}
\arguments{
\item{x}{object of class Survey.Design}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Summarises an S4 object of class 'Survey.Design'
}
